// projects/codx-mwp/src/lib/model/Storages.model.ts
export class Storages {
    recID: string;
    title: string;
    memo: string;
    storageType: string;
    details: any[];
    createdOn: string;
    createdBy: string;   
    modifiedOn: string;
}

export class StorageGoal {
    recID: string;
    refID: string;
    memo: string;
    createdOn: string;
    createdBy: string;   
}